# web/schedule_api.py
from flask import Blueprint, jsonify, request, send_file
from flask_limiter import Limiter
from flask_limiter.util import get_remote_address
import requests
from config import API_URL
import os
from datetime import datetime
from concurrent.futures import ThreadPoolExecutor
from telegram_auth import authenticated_telegram_user

schedule_bp = Blueprint('schedule', __name__)
limiter = Limiter(
    app=None,
    key_func=get_remote_address,
    default_limits=["10 per second"]
)

def _get_authenticated_user_id():
    telegram_user = authenticated_telegram_user(request)
    return telegram_user['id'] if telegram_user else None


def _json_from_upstream(response):
    try:
        return response.json()
    except ValueError:
        return {"error": "Ошибка сервера"}


def _public_user_payload(payload):
    if not isinstance(payload, dict):
        return payload
    role = payload.get("role")
    if role in {"Я преподаватель", "teacher"}:
        payload["role"] = "teacher"
    elif role in {"Я студент", "student"}:
        payload["role"] = "student"
    return payload


_CATALOG_KINDS = ("groups", "teachers", "audiences")


def _fetch_catalog_values(kind):
    response = requests.get(f"{API_URL}/{kind}", timeout=5)
    if response.status_code == 404:
        return []
    response.raise_for_status()
    payload = _json_from_upstream(response)
    values = payload.get(kind, []) if isinstance(payload, dict) else []
    if not isinstance(values, list):
        return []
    return [value.strip() for value in values if isinstance(value, str) and value.strip()]


def _catalog_json_response(payload):
    response = jsonify(payload)
    response.headers['Cache-Control'] = 'private, max-age=60, stale-while-revalidate=300'
    return response


def _single_catalog_response(kind):
    try:
        values = _fetch_catalog_values(kind)
    except requests.RequestException:
        values = []
    return _catalog_json_response({kind: values})


@schedule_bp.route('/api/schedule/catalog')
@limiter.limit("10 per second")
def get_schedule_catalog():
    """Return current-academic-year entities in one mobile-friendly request."""
    with ThreadPoolExecutor(max_workers=len(_CATALOG_KINDS)) as executor:
        futures = {
            kind: executor.submit(_fetch_catalog_values, kind)
            for kind in _CATALOG_KINDS
        }
        payload = {}
        for kind, future in futures.items():
            try:
                payload[kind] = future.result()
            except requests.RequestException:
                payload[kind] = []
    return _catalog_json_response(payload)


# Keep old clients quiet while cached schedule-search assets expire.
@schedule_bp.route('/api/groups')
@limiter.limit("10 per second")
def get_groups_catalog():
    return _single_catalog_response('groups')


@schedule_bp.route('/api/teachers')
@limiter.limit("10 per second")
def get_teachers_catalog():
    return _single_catalog_response('teachers')


@schedule_bp.route('/api/audiences')
@limiter.limit("10 per second")
def get_audiences_catalog():
    return _single_catalog_response('audiences')


@schedule_bp.route('/api/me')
@limiter.limit("10 per second")
def get_me():
    """
    Получение данных авторизованного пользователя
    ---
    tags:
      - Расписание
    parameters:
    responses:
      200:
        description: Данные пользователя успешно получены
        schema:
          type: object
          properties:
            user_id:
              type: integer
              description: ID пользователя
            role:
              type: string
              description: Роль пользователя
            name_or_group:
              type: string
              description: Имя или группа пользователя
            is_class_teacher:
              type: boolean
              description: Является ли классным руководителем
            class_group:
              type: string
              description: Классная группа
      404:
        description: Пользователь не найден
      500:
        description: Ошибка сервера
    """
    try:
        user_id = _get_authenticated_user_id()
        if not user_id:
            return jsonify({"error": "Требуется авторизация Telegram"}), 401
        response = requests.get(f"{API_URL}/user/{user_id}", timeout=5)
        return jsonify(_public_user_payload(_json_from_upstream(response))), response.status_code
    except requests.RequestException:
        return jsonify({"error": "Ошибка сервера"}), 500


@schedule_bp.route('/api/schedule/me', methods=['POST'])
@limiter.limit("10 per second")
def get_my_schedule():
    """
    Получение расписания авторизованного пользователя
    ---
    tags:
      - Расписание
    parameters:
      - name: body
        in: body
        required: true
        schema:
          type: object
          required: ["date"]
          properties:
            date:
              type: string
              pattern: "^\\d{2}\\.\\d{2}\\.\\d{4}$"
              description: Дата в формате DD.MM.YYYY
              example: "22.04.2025"
    responses:
      200:
        description: Расписание успешно получено
        schema:
          type: object
          properties:
            group:
              type: string
              description: Название группы
            teacher:
              type: string
              description: Имя преподавателя
            date:
              type: string
              description: Дата расписания
            schedule:
              type: array
              items:
                type: object
                properties:
                  time:
                    type: string
                    description: Время занятия
                  subject:
                    type: string
                    description: Название предмета
                  room:
                    type: string
                    description: Номер аудитории
      400:
        description: Неверный формат даты (должен быть DD.MM.YYYY) или отсутствует параметр date
      404:
        description: Пользователь не найден
      500:
        description: Ошибка сервера
    """
    try:
        user_id = _get_authenticated_user_id()
        if not user_id:
            return jsonify({"error": "Требуется авторизация Telegram"}), 401
        response = requests.post(
            f"{API_URL}/getUserSchedule/{user_id}",
            json=request.json,
            headers={'Content-Type': 'application/json'},
            timeout=10,
        )
        return jsonify(_json_from_upstream(response)), response.status_code
    except requests.RequestException:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/schedule/group', methods=['POST'])
@limiter.limit("10 per second")
def get_group_schedule():
    """
    Получение расписания группы
    ---
    tags:
      - Расписание
    parameters:
      - name: body
        in: body
        required: true
        schema:
          type: object
          required: ["group"]
          properties:
            group:
              type: string
              description: Название группы
            date:
              type: string
              pattern: "^\\d{2}\\.\\d{2}\\.\\d{4}$",
              description: Дата в формате DD.MM.YYYY
              example: "22.04.2025"
    responses:
      200:
        description: Расписание группы успешно получено
      404:
        description: Группа не найдена в расписании
      500:
        description: Ошибка сервера
    """
    try:
        response = requests.post(
            f"{API_URL}/schedule/group",
            json=request.json,
            headers={'Content-Type': 'application/json'}
        )
        return jsonify(response.json()), response.status_code
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/schedule/teacher', methods=['POST'])
@limiter.limit("10 per second")
def get_teacher_schedule():
    """
    Получение расписания преподавателя
    ---
    tags:
      - Расписание
    parameters:
      - name: body
        in: body
        required: true
        schema:
          type: object
          required: ["teacher"]
          properties:
            teacher:
              type: string
              description: Имя преподавателя
            date:
              type: string
              pattern: "^\\d{2}\\.\\d{2}\\.\\d{4}$",
              description: Дата в формате DD.MM.YYYY
              example: "22.04.2025"
    responses:
      200:
        description: Расписание преподавателя успешно получено
      404:
        description: Преподаватель не найден в расписании
      500:
        description: Ошибка сервера
    """
    try:
        response = requests.post(
            f"{API_URL}/schedule/teacher",
            json=request.json,
            headers={'Content-Type': 'application/json'}
        )
        return jsonify(response.json()), response.status_code
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/schedule/audience', methods=['POST'])
@limiter.limit("10 per second")
def get_audience_schedule():
    """
    Получение расписания аудитории
    ---
    tags:
      - Расписание
    parameters:
      - name: body
        in: body
        required: true
        schema:
          type: object
          required: ["audience"]
          properties:
            audience:
              type: string
              description: Номер аудитории
            date:
              type: string
              pattern: "^\\d{2}\\.\\d{2}\\.\\d{4}$",
              description: Дата в формате DD.MM.YYYY
              example: "22.04.2025"
    responses:
      200:
        description: Расписание аудитории успешно получено
      404:
        description: Аудитория не найдена в расписании
      500:
        description: Ошибка сервера
    """
    try:
        response = requests.post(
            f"{API_URL}/schedule/audience",
            json=request.json,
            headers={'Content-Type': 'application/json'}
        )
        return jsonify(response.json()), response.status_code
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/schedule/download/<schedule_type>/<date>')
@limiter.limit("10 per second")
def download_schedule(schedule_type, date):
    """
    Скачивание файла расписания
    ---
    tags:
      - Расписание
    parameters:
      - name: schedule_type
        in: path
        type: string
        required: true
        enum: [groups, teachers]
        description: Тип расписания
      - name: date
        in: path
        type: string
        required: true
        pattern: "^\\d{2}\\.\\d{2}\\.\\d{4}$"
        description: Дата в формате DD.MM.YYYY
    responses:
      200:
        description: Файл расписания успешно скачан
        schema:
          type: file
      400:
        description: Неверный тип расписания
      404:
        description: Файл не найден
      500:
        description: Ошибка сервера
    """
    try:
        if schedule_type not in ["groups", "teachers"]:
            return jsonify({"error": "Неверный тип расписания"}), 400

        response = requests.get(
            f"{API_URL}/schedule/download/{schedule_type}/{date}",
            stream=True
        )

        if response.status_code == 200:
            # Определяем тип и имя файла из ответа FastAPI
            content_type = response.headers.get('content-type', 'application/octet-stream')
            content_disp = response.headers.get('content-disposition', '')

            # Извлекаем оригинальное имя файла из content-disposition
            import re as _re
            from urllib.parse import unquote
            fname_match = _re.search(r"filename\*=utf-8''(.+)", content_disp)
            if fname_match:
                download_name = unquote(fname_match.group(1))
            else:
                ext = '.pdf' if 'pdf' in content_type else '.xlsx'
                download_name = f"schedule_{schedule_type}_{date}{ext}"

            return send_file(
                response.raw,
                mimetype=content_type,
                as_attachment=True,
                download_name=download_name
            )
        else:
            return jsonify({"error": "Файл не найден"}), 404
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/groups')
@limiter.limit("10 per second")
def get_all_groups():
    """
    Получение списка всех групп
    ---
    tags:
      - Расписание
    responses:
      200:
        description: Список групп успешно получен
        schema:
          type: object
          properties:
            groups:
              type: array
              items:
                type: string
                description: Название группы
      404:
        description: Список групп не найден
      500:
        description: Ошибка сервера
    """
    try:
        response = requests.get(f"{API_URL}/groups")
        return jsonify(response.json()), response.status_code
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500

@schedule_bp.route('/api/teachers')
@limiter.limit("10 per second")
def get_all_teachers():
    """
    Получение списка всех преподавателей
    ---
    tags:
      - Расписание
    responses:
      200:
        description: Список преподавателей успешно получен
        schema:
          type: object
          properties:
            teachers:
              type: array
              items:
                type: string
                description: ФИО преподавателя
      404:
        description: Список преподавателей не найден
      500:
        description: Ошибка сервера
    """
    try:
        response = requests.get(f"{API_URL}/teachers")
        return jsonify(response.json()), response.status_code
    except Exception as e:
        return jsonify({"error": "Ошибка сервера"}), 500


@schedule_bp.route('/api/audiences')
@limiter.limit("10 per second")
def get_all_audiences():
    """Получение списка аудиторий текущего учебного года."""
    try:
        response = requests.get(f"{API_URL}/audiences", timeout=10)
        return jsonify(_json_from_upstream(response)), response.status_code
    except requests.RequestException:
        return jsonify({"error": "Ошибка сервера"}), 500